% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_edit.R
\name{data_edit}
\alias{data_edit}
\title{An interactive editor for viewing, entering and editing data}
\usage{
data_edit(
  x = NULL,
  col_bind = NULL,
  col_edit = TRUE,
  col_options = NULL,
  col_stretch = FALSE,
  col_factor = FALSE,
  col_names = TRUE,
  col_readonly = NULL,
  col_hide = NULL,
  row_bind = NULL,
  row_edit = TRUE,
  save_as = NULL,
  title = NULL,
  logo = NULL,
  logo_size = 30,
  logo_side = "left",
  viewer = "dialog",
  viewer_height = 800,
  viewer_width = 1200,
  theme = "yeti",
  read_fun = "read.csv",
  read_args = NULL,
  write_fun = "write.csv",
  write_args = NULL,
  quiet = FALSE,
  hide = FALSE,
  code = FALSE,
  cancel,
  track = NULL,
  ...
)
}
\arguments{
\item{x}{a matrix, data.frame, data.table or the name of a csv file to edit.
Tibbles are also supported but will be coerced to data.frames. An empty
table can be created by specifying the dimensions in a vector of the form
\code{c(nrow, ncol)} or the names of the columns to include in the
template.}

\item{col_bind}{additional columns to add to the data prior to loading into
editor, can be either an array containing the new data, a vector containing
the new column names for empty columns or a named list containing a vector
for each new column.}

\item{col_edit}{logical indicating whether columns can be added or removed,
set to TRUE by default.}

\item{col_options}{named list containing the options for columns that use
dropdown menus, dates, checkboxes or passwords.}

\item{col_stretch}{logical indicating whether columns should be stretched to
fill the full width of the display, set to FALSE by default.}

\item{col_factor}{logical indicating whether character columns should be
converted to factors prior to returning the edited data, set to FALSE by
default.}

\item{col_names}{logical indicating whether column names can be edited or a
vector of column names that cannot be edited, set to TRUE by default to
allow editing of column names.}

\item{col_readonly}{names of columns that cannot be edited. Users will be
able to edit values but these will be reverted to the original values.
Column names for these column cannot be edited either.}

\item{col_hide}{names of columns to hide from the editor. Hidden columns will
not be visible or editable but will be retained in the returned data.}

\item{row_bind}{additional rows to add to the data prior to loading into
editor, can be either an array containing the new data, a vector containing
the new row names for empty rows or a named list containing a vector for
each new column.}

\item{row_edit}{logical indicating whether rows can be added or removed, set
to TRUE by default.}

\item{save_as}{name of a csv file to which the edited data should be saved.}

\item{title}{optional title to include above the data editor.}

\item{logo}{optional package logo to include in title above the data editor,
must be supplied as path to logo png.}

\item{logo_size}{width of the logo in pixels, set to 30 pixels by default.}

\item{logo_side}{can be either \code{"left"} or \code{"right"} to determine
the position of the logo relative to the title, set to \code{"left"} by
default.}

\item{viewer}{can be either \code{"dialog"}, \code{"browser"} or
\code{"pane"} to open the application in a dialog box, browser or RStudio
viewer pane. First letter abbreviations are allowed, set to \code{"dialog"}
by default.}

\item{viewer_height}{numeric to control the height of the viewer in pixels
when \code{viewer} is set to \code{"dialog"}, set 800 by default.}

\item{viewer_width}{numeric to control the width of the viewer in pixels when
\code{viewer} is set to \code{"dialog"}, set to 1200 by default.}

\item{theme}{valid shinytheme name, set to "yeti" by default.}

\item{read_fun}{name of the function to use to read in the data when \code{x}
is the name of a file, set to \code{read.csv} by default.}

\item{read_args}{a named list of additional arguments to pass to
\code{read_fun}.}

\item{write_fun}{name of the function to use to write the edited version of
\code{x} to a file, set to \code{write.csv} by default. Only requirement is
that the first argument accepts the edited data and the second argument
accepts the file name supplied to \code{save_as}.}

\item{write_args}{a named list of additional arguments to pass to
\code{write_fun}.}

\item{quiet}{logical indicating whether messages should be suppressed, set to
FALSE by default.}

\item{hide}{logical indicating whether the \code{dataInput} and
\code{dataOutput} modules should be visible to the user within the
application. If \code{hide = FALSE} and \code{save_as} is specified, the
edited data will be written to file after the application is closed.}

\item{code}{logical indicating whether tidyverse code required to replicate
the data edits should be printed to the console, set to \code{FALSE} by
default. Alternatively, users can supply the name of an R script to write
the code to. The generated code uses \code{dplyr} verbs such as
\code{rename()}, \code{select()}, \code{mutate()}, \code{slice()}, and
\code{tibble::add_row()} to replicate column renames, column
additions/removals, cell value changes, and row additions/removals.}

\item{cancel}{optional value to return when the user hits the \code{cancel}
button, set to the supplied data by default.}

\item{track}{can be set to \code{TRUE} to highlight cells that have been
edited or added to the original data with a default blue border, or a valid
CSS color (e.g. \code{"#FF0000"} or \code{"red"}) to use a custom border
color. Set to \code{NULL} by default to disable highlighting.}

\item{...}{not in use.}
}
\value{
the edited data as a matrix or data.frame.
}
\description{
\code{data_edit} is a shiny application built on \code{rhandsontable} that is
designed to make it easy to interactively view, enter or edit data without
any coding. \code{data_edit} is also a wrapper for any reading or writing
function to make it easy to interactively update data saved to file.
}
\examples{
if(interactive()) {

  data_edit(mtcars)

}

}
\author{
Dillon Hammill, \email{dillon.hammill21@gmail.com}
}
