% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFilter.R
\name{dataFilter}
\alias{dataFilter}
\alias{dataFilterUI}
\alias{dataFilterServer}
\title{Shiny module for filtering data}
\usage{
dataFilterUI(id)

dataFilterServer(id, data = reactive(NULL))
}
\arguments{
\item{id}{unique identifier for the module to prevent namespace clashes when
making multiple calls to this shiny module.}

\item{data}{an array wrapped in \code{reactive()} containing the data to be
filtered.}
}
\value{
a list of reactive objects containing the filtered \code{data} and
indices for filtered \code{rows}.
}
\description{
Shiny module for filtering data
}
\examples{
if (interactive()) {
  library(shiny)
  library(rhandsontable)
  library(shinyjs)

  ui <- fluidPage(
    useShinyjs(),
    dataInputUI("input1"),
    dataFilterUI("filter1"),
    rHandsontableOutput("data1")
  )

  server <- function(input,
                     output,
                     session) {
    data_input <- dataInputServer("input1")

    data_filter <- dataFilterServer("filter1",
      data = data_input
    )

    output$data1 <- renderRHandsontable({
      if (!is.null(data_filter())) {
        rhandsontable(data_filter())
      }
    })
  }

  shinyApp(ui, server)
}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
