% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{dataOutput}
\alias{dataOutput}
\alias{dataOutputUI}
\alias{dataOutputServer}
\title{Shiny module for data output}
\usage{
dataOutputUI(id, cellWidths = c("60\%", "38\%"))

dataOutputServer(
  id,
  data = reactive(NULL),
  save_as = NULL,
  write_fun = "write.csv",
  write_args = NULL,
  ...
)
}
\arguments{
\item{id}{unique identifier for the module to prevent namespace clashes when
making multiple calls to this shiny module.}

\item{cellWidths}{a vector of length 2 to control the relative widths of the
\code{textInput} and \code{actionButton}, set to \code{c("60\%", "38\%")}
by default to optimise user interface for \code{\link{data_edit}}.}

\item{data}{an object of class data.frame wrapped in \code{reactive} to be
saved to file.}
}
\description{
Shiny module for data output
}
\examples{
if(interactive()) {

  library(shiny)
  library(rhandsontable)

  ui <- fluidPage(
    dataInputUI("input1"),
    dataOutputUI("output1"),
    rHandsontableOutput("data1")
  )

  server <- function(input,
                     output,
                     session) {

    data_input1 <- dataInputServer("input1")

    output$data1 <- renderRHandsontable({
      if(!is.null(data_input1())) {
        rhandsontable(data_input1())
      }
    })
    
    dataOutputServer("output1",
                     data = data_input1)

  }
  
  shinyApp(ui, server)

}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
