% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{dataSelect}
\alias{dataSelect}
\alias{dataSelectUI}
\alias{dataSelectServer}
\title{Shiny module for selecting data}
\usage{
dataSelectUI(id)

dataSelectServer(id, data = reactive(NULL))
}
\arguments{
\item{id}{unique identifier for the module to prevent namespace clashes when
making multiple calls to this shiny module.}

\item{data}{an array wrapped in \code{reactive()} containing the data to be
filtered.}
}
\description{
Shiny module for selecting data
}
\examples{
if(interactive()) {

  library(shiny)
  library(rhandsontable)
  library(shinyjs)

  ui <- fluidPage(
    useShinyjs(),
    dataInputUI("input1"),
    dataSelectUI("select1"),
    dataOutputUI("output1"),
    rHandsontableOutput("data1")
  )

  server <- function(input,
                     output,
                     session) {

    data_input <- dataInputServer("input1")

    data_select <- dataSelectServer("select1",
    data = data_input)

    output$data1 <- renderRHandsontable({
      if(!is.null(data_select())) {
        rhandsontable(data_select())
      }
    })

    dataOutputServer("output1",
                     data = data_select)

  }

  shinyApp(ui, server)

}

}
\author{
Dillon Hammill, \email{Dillon.Hammill@anu.edu.au}
}
